refactor debt calendar
